totem/python/devices/eink/pyproject.toml
[tool.poetry]
name = "eink-device"
version = "0.1.0"
description = "E-Ink screen drivers for Totem"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "devices" }]

[tool.poetry.dependencies]
python = "^3.9"
pillow = ">=9.2.0,<12.0.0"
numpy = "^1.23.0"
RPi-GPIO = { version = "^0.7.0", optional = true, markers = "sys_platform == 'linux'" }
spidev = { version = "^3.5", optional = true, markers = "sys_platform == 'linux'" }

[tool.poetry.extras]
hardware = ["RPi-GPIO", "spidev"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"